backend/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
gunicorn
redis
celery
python-dotenv

djangorestframework-simplejwt>=5.3
django-cors-headers
django-filter
Pillow
